SDA-8775: Align label and taints validation between create and edit
