+++ pkg/helper/machinepools/helpers_suite_test.go
@@ -1,0 +1,13 @@
+package machinepools
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestIdp(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "MachinePools Suite")
+}
